+++ src/main/java/com/remedios/lucas/curso/alunoDisciplina/DadosCadastroAlunoDisciplina.java
@@ -3,9 +3,9 @@ package com.remedios.lucas.curso.alunoDisciplina;
 public record DadosCadastroAlunoDisciplina(
         Long idAluno,
         Long idDisciplina,
-        Long nota1,
-        Long nota2,
-        Long notaAtividade,
-        Long media
+        Double nota1,
+        Double nota2,
+        Double notaAtividade,
+        Double media
 ) {
 }
